Rename bidx as index
