+++ index/worker.go
@@ -17,7 +17,7 @@
 // TODO(jiawei)
 // This is supposed to send out RPC calls if server instance does not handle
 // given attribute / predicate.
-package bidx
+package index
 
 import (
 	"log"
